-New module: battery module for /sys BAT0. Flexibility to come.
-Plugged memory leak in moc module.

-Plugged similar perceived potential memory leaks in cpuinfo module.
Also reenabled cpuname and cpucache.

-Fixed issue with timeout 0 modules not properly redrawing when
displaced.

-Added old_pixel_width to text_section struct to help with more efficient use
of clear_area() when we get the damn thing to work.

-clear_area() would have to be called before we set old_xpos = new xpos for
the next loop. It should also only be called if old_xpos +
old_pixel_width is greater than the new xpos + new pixel_width. If
that's not the case, there's nothing to clear.

-Changed xcb_clear_area() to xcb_clear_area_checked() in our clear_area().

-Static text pixel_width only needs to be set once.