Nothing that big... small tweaks here and there. Some out of sheer boredom.
Tweaked the MPD module to (hopefully) be using more efficient standard
lib functions. Donky still appears to be using more CPU than I would
like (not that it's significant to begin with). It probably isn't even
the MPD module, like I thought it might've been.

Added a tiny event poll to redraw donky after it's been obscured in a
way that isn't very sophisticated, but hey, it works.