Added support for VARIABLE_BAR.
Overall, i'm happy with the result, only thing needs to be tweaked is
the positioning on the line.  Right now it places it at the top of the
line, should probably find a different method.  I'm too burned out to
wrap my mind around anything else.  The foundation is totally layed out
though, just minor tweaking.

I added an mpd_volume_bar variable to the mpd module.  This works great,
and was a good first test.  Need to implement a progress bar also.