Changed the get_int_key(), get_char_key(), etc. functions to receive an "otherwise" parameter which it will return if it doesn't find the requested key. This automates/largely removes the need to check the return results on these functions. This cleaned up some code, especially in x11.c where we need it. Updated everything that used these functions.
Added a couple more things into default_settings.h.

Changed freeif() to NULL the pointer passed to it after freeing.

For some reason, I can't get freeif() to work correctly in the battery
module. This is unrelated to the change. Also, I probably changed other
stuff in the battery module.