donky is now release under the CC0 and is therefore for all intents and purposes public domain software where possible (all copyrights have been waived). Where not possible, the CC0 essentially acts as an absolutely permissable license with absolutely no strings attached. By no strings attached, I mean things such as attribution are not required by the CC0 fallback license.
Personally, I urge you to NOT give us attribution. The code is free,
period, so just leave it at that. There's no need to go polluting your
source files with attributions and legal bullshit. We hate licenses and
don't feel like imposing ANY obligations on anyone, including requiring
people to put lame ass copyright notices at the top of files.

See the COPYING file for a copy of the CC0 and complete information.

Alternatively, see these websites:
http://creativecommons.org/publicdomain/zero/1.0/
http://creativecommons.org/publicdomain/zero/1.0/legalcode

To apply the CC0 to your own work (I strongly encourage you to, and it
certainly doesn't have to be software), please see:
http://creativecommons.org/license/zero/